MemoryEditor: Clean and reorganization to abstract out t_memory_range concept for use by other tools. CheatFinder: Added empty skeleton.

+++ branches/20110530-allegro5-sound/srcs/smsid/replacement_code.cpp
@@ -9,8 +9,8 @@
 // from meka.c
 t_meka_configuration	g_configuration;
 
-// from memory.c
-void* Memory_Alloc(int size)
+// from misc.c
+void* Memory_Alloc(size_t size)
 {
 	return malloc(size);
 }
